+++ src/NHA/NHA.php
@@ -66,7 +66,7 @@ use NHA\Repository\WorldRepository;
  * @property SocialRepository         $social
  * @property WorldRepository          $world
  *
- * @version 3.15.0
+ * @version 3.16.0
  */
 class NHA extends MessageCommandClient
 {